requirements.txt
eventlet
msgpack-python
